requirements.txt
appdata==2.1.2
asgiref==3.5.2
click==8.0.3
colored==1.4.3
Django==4.0.6
django-froala-editor==4.0.13
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
to-requirements.txt==1.3.0
tzdata==2022.1
